[FEAT] 实现AT24C256驱动程序

+++ Demo/04.串口/03.封装printf打印函数/User/main.c
@@ -9,7 +9,7 @@ int main(void)
 
     /* 验证下来发现 %x 无法正确打印，其他格式没有问题 */
     /* 打印结果为: H 1000 demo 5500 Hello World! */
-    printf("%c %d %s %x Hello World!\r\n", 'H', 1000, "demo", 0x55);
+    printf("%c %d %s %x Hello World!\r\n", 'H', (int)1000, "demo", (int)0x55);
 
     while(1);
 }
